ai/llm/requirements.txt
fastapi
uvicorn
python-multipart
faster-whisper
httpx
python-dotenv
langchain
langchain-openai
langchain-anthropic

# HTTP client for keep-alive service
aiohttp

# LangSmith for monitoring and analytics
langsmith

# RabbitMQ
pika==1.3.2

# ChromaDB and embeddings
chromadb==1.0.15
sentence-transformers==5.0.0

# Data processing
pandas==2.3.1
numpy==2.0.2

# MinIO client
minio==7.2.0

# PostgreSQL dependencies
asyncpg==0.29.0              # 비동기 PostgreSQL 어댑터
psycopg2-binary==2.9.9       # 동기 PostgreSQL 어댑터 (RabbitMQ service용)
sqlalchemy[asyncio]==2.0.23
